Issue #11: Reorganized views into folder module (`show_registration_form_success` moved to `helpers`).
